+++ apps/marmicode-e2e/src/support/e2e.ts
@@ -1,2 +1,2 @@
-import '@marmicode/testing/cypress-commands';
+import '@marmicode/shared/testing/cypress-commands';
 import '@jscutlery/cypress-harness/support';
