+++ tsconfig.base.json
@@ -49,12 +49,12 @@
       ],
       "@marmicode/shared/ui": ["libs/shared/ui/index.ts"],
       "@marmicode/shared/utils": ["libs/shared/utils/index.ts"],
-      "@marmicode/testing": ["libs/testing/src/index.ts"],
-      "@marmicode/testing/cypress-commands": [
-        "libs/testing/src/cypress-commands.ts"
+      "@marmicode/shared/testing": ["libs/shared/testing/src/index.ts"],
+      "@marmicode/shared/testing/cypress-commands": [
+        "libs/shared/testing/src/cypress-commands.ts"
       ],
-      "@marmicode/testing/cypress-ct-commands": [
-        "libs/testing/src/cypress-ct-commands.ts"
+      "@marmicode/shared/testing/cypress-ct-commands": [
+        "libs/shared/testing/src/cypress-ct-commands.ts"
       ],
       "@marmicode/tools/build-github-workflows": [
         "tools/build-github-workflows/src/index.ts"
